Feature: Display incomplete page
There are 3 files: todo.js, dom-creator.js and form-creator.js.

todo.js just creates a todo object. It has title, description, due date
and priority at the moment.

dom-creator.js makes the elements and loads them on the html. It uses
the FormCreator class to make the form, so the code is divided in two
files rather than only one.

form-creator.js contains a class and functions to make the form. It also
adds a listener to the form, thanks to event delegation to handle logic
when the user clicks the cancel button or the submit button.

Added a line to webpack config to help debugging during development.