Add: create project form
Added a button that creates a form that asks for a new project name.

Refactor both dom-creator.js and form-creator.js. In dom-creator.js,
there is a function to create both forms and another function to give
them logic when the buttons are clicked. Also moved functions.

In form-creator.js, there is another function to create the project
form. Changed and moved some functions to make both the createToDoForm
and the project form similar.