Feature: Add logic to create new projects
For the moment, a user can use the form to create a project and it will
be displayed on the list. Also, there will be a container with a class
name that corresponds with the project name.

There is some refactoring too. There are new files, like a page
controller to handle objects communicating between each other, a project
manager, to handle the projects. Also there is a todo manager, but it's
not functional at the moment.

There are also some new functions in the previous files to avoid
repetition. And replaced some parts of code where it explicitly calls a
new statement to create an object. Now in the constructor, it has a
parameter that expects an object, with the object used as a default.