Feature: display created todos
Now the user can see the created todos. For the moment, it can only see
those that correspond to the selected project (default when not
creating a new one).

In order to do this, many functions from event-listener-agregator now
accept an args parameter. This args parameter is an object that contains
other objects; between these objects are the projecManager object,
needed to obtain the current project and add it.

There are also new methods that create the necessary elements to display
the new todo. This todo is stored in a li, with a button and a paragraph
element that shows its title and description. The button doesn't do
anything at the moment. In the future, when clicked, the todo will be
marked as complete.

Also another thing to implement is the ability to modify or even delete
any todo displayed.