Feature: let the user switch between projects
The user can now change between the projects. These projects shown in
the sidebar now are buttons.

In page-controller.js there is a new method that returns an object with
the params sent to addNavbarListener. There is also another call to
other listener method.

In event-listener-agregator.js, it adds another listener through a new
method, now to the projects buttons. It ensures that when the user
clicks one of the buttons, it calls the same method to replace the
project contents on the page.

In dom-creator.js, there are a few changes to some method parameters.
Now they accept args, like other methods, args is an object with other
objects to access another methods. This file now also has new methods
to check if a user has created todos in a project. When that's the case,
it obtains the project object, then its todos, and calls a new method
to make them li objects and pass them to the other methods used to load
the content.