Feature: let users expand and collapse todos
When a user clicks a todo, the first click will "expand" the todo. This
means that the todo will show all the remaining information. And when
the user clicks again the same todo, it will "collapse". This means that
it will return to when the todo was created.

There are new methods in dom-creator.js. One of them removes an element
passed. Another one passes a query and returns the result of said query.
There's also getter/setter methods in this file.

In event-listener-agregator.js now has 3 methods. All of them just add
a listener to some element. The previous methods that added logic to
elements have been moved to page-controller.js.

In todo-manager.js, the method that gathers data now returns an object.
Also created another method to properly validate data. Also changed the
way it displays the objects in this file, to use many lines.

In project.js, there is now a method to search a todo, used in the new
feature.

In project-manajer.js, refactor some methods.

In page-controller.js, as previously said, now contains methods to
handle logic when the user clicks certain elements. Maybe these methods
will be moved later to another module. There is also a method that
creates a form creator instance variable, used in one of the new
methods. Also, it contains methods to expand and collapse the todos.
Finally, it adds a line to keep adding the listener to the current
project when changed.