Feature: add logic to delete todos
The user can now delete a todo using a new button. This button searches
the todo object and with new methods it deletes it from the project
where it's stored. It also deletes the li object.

Fixed a bug where if the user creates a new project, this new project
didn't have the listener that handles the todos and its buttons.