Feature: Edit todos
The user can now edit todos.

To let the user do this action, the button named mark now its named edit
and added a listener that opens the form previously used to create a new
todo. When the user clicks the submit button of said form, it will
update the todo object properties first, then update the todo object
stored in its project and finally it updates the info on the page.

For that there are a few new methods in some modules to achieve this
behavior.

There is a "bug" where if the user expands a todo and then edits it, the
expanded information doesn't update properly, it needs to make two
clicks to fix it.