Fix: update expanded todo when edited
The todo now updates properly.

In the previous commit, there is a notice about a bug that doesn't
update the todo when edited. A manual fix was to click two times the
todo, to collapse it and then expand it again.

To fix it, the expand and collapse methods located in page-controller
are now in dom-creator, now being easy to be called from the
todo-manager because it has access to the dom-creator object.

Also, the handleToDo method is refactored to reflect that the condition
to expand or collapse is based on the amount of paragraphs in the li
element. There is a new method in dom-creator to do this.