Feature: Add localStorage
When refreshing the page, the projects and todos are deleted. Using the
WebStorage API (localStorage specifically) solves this problem.

Now the page-controller.js contains methods that deal with this feature.
Also there are methods in other files to help achieve it.

There was a bug when the user creates a project, it displayed an error
in the console, trying to delete a form that was actually deleted before
with another method. So to avoid this, now it doesn't call the method
that deletes a form.

Currently there is a bug where if the user has two or more projects, the
default project button in the sidebar goes to the bottom of this list of
buttons. When the user loads the page for the first time, this default
button is the first in the list.