Fix: todos now expand/collapse correctly
Before, to expand/collapse a todo, you need to click below the last
paragraph element or in the li itself. If clicking to the right of the
buttons, it will detect a div that contains the form buttons and that
prevented to access the data needed to this behavior.

The solution now checks if the element clicked is the li and if not, it
gets the li using a recursive function.